# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teenagers Choir API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  message: string;
}

export interface LoginBody {
  username: string;
  password: string;
}

export interface RegisterBody {
  username: string;
  password: string;
}

export interface ChangePasswordBody {
  currentPassword: string;
  newPassword: string;
}

export interface AuthUser {
  id: number;
  username: string;
  role: string;
  voicePart: string;
}

export interface User {
  id: number;
  username: string;
  role: string;
  voicePart: string;
  createdAt: string;
}

export type UpdateVoicePartBodyVoicePart =
  (typeof UpdateVoicePartBodyVoicePart)[keyof typeof UpdateVoicePartBodyVoicePart];

export const UpdateVoicePartBodyVoicePart = {
  Soprano: "Soprano",
  Alto: "Alto",
  Normal: "Normal",
} as const;

export interface UpdateVoicePartBody {
  voicePart: UpdateVoicePartBodyVoicePart;
}

export type AttendanceRecordStatus =
  (typeof AttendanceRecordStatus)[keyof typeof AttendanceRecordStatus];

export const AttendanceRecordStatus = {
  present: "present",
  absent: "absent",
} as const;

export interface AttendanceRecord {
  id: number;
  userId: number;
  username: string;
  voicePart: string;
  status: AttendanceRecordStatus;
  date: string;
}

export type UpdateAttendanceBodyStatus =
  (typeof UpdateAttendanceBodyStatus)[keyof typeof UpdateAttendanceBodyStatus];

export const UpdateAttendanceBodyStatus = {
  present: "present",
  absent: "absent",
} as const;

export interface UpdateAttendanceBody {
  status: UpdateAttendanceBodyStatus;
}

export interface AttendanceStats {
  totalSessions: number;
  presentCount: number;
  absentCount: number;
  attendanceRate: number;
}

export interface Message {
  id: number;
  content: string;
  /** @nullable */
  targetVoicePart: string | null;
  createdAt: string;
  authorUsername: string;
}

export interface CreateMessageBody {
  content: string;
  /** @nullable */
  targetVoicePart: string | null;
}

export type MusicFileFileType =
  (typeof MusicFileFileType)[keyof typeof MusicFileFileType];

export const MusicFileFileType = {
  pdf: "pdf",
  mp3: "mp3",
  other: "other",
} as const;

export interface MusicFile {
  id: number;
  title: string;
  url: string;
  fileType: MusicFileFileType;
  /** @nullable */
  targetVoicePart: string | null;
  createdAt: string;
}

export type CreateMusicBodyFileType =
  (typeof CreateMusicBodyFileType)[keyof typeof CreateMusicBodyFileType];

export const CreateMusicBodyFileType = {
  pdf: "pdf",
  mp3: "mp3",
  other: "other",
} as const;

export interface CreateMusicBody {
  title: string;
  url: string;
  fileType: CreateMusicBodyFileType;
  /** @nullable */
  targetVoicePart: string | null;
}

export interface AdminStats {
  totalMembers: number;
  sopranoCount: number;
  altoCount: number;
  normalCount: number;
  presentToday: number;
  absentToday: number;
  totalMessages: number;
  totalMusicFiles: number;
}