# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Teenagers Choir API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /messages:
    get:
      operationId: listMessages
      tags: [messages]
      summary: Get messages for the current user's voice part
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: createMessage
      tags: [messages]
      summary: Send a broadcast message (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMessageBody"
      responses:
        "201":
          description: Created message
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
  /messages/{id}:
    delete:
      operationId: deleteMessage
      tags: [messages]
      summary: Delete a message (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /music:
    get:
      operationId: listMusic
      tags: [music]
      summary: Get music files for current user's voice part
      responses:
        "200":
          description: Music files
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MusicFile"
    post:
      operationId: createMusic
      tags: [music]
      summary: Upload a music link (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMusicBody"
      responses:
        "201":
          description: Created music entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MusicFile"
  /music/{id}:
    delete:
      operationId: deleteMusic
      tags: [music]
      summary: Delete a music file (admin only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    Message:
      type: object
      properties:
        id:
          type: integer
        content:
          type: string
        targetVoicePart:
          type: ["string", "null"]
        createdAt:
          type: string
        authorUsername:
          type: string
      required:
        - id
        - content
        - targetVoicePart
        - createdAt
        - authorUsername
    CreateMessageBody:
      type: object
      properties:
        content:
          type: string
        targetVoicePart:
          type: ["string", "null"]
      required:
        - content
        - targetVoicePart
    MusicFile:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        url:
          type: string
        fileType:
          type: string
          enum: [pdf, mp3, other]
        targetVoicePart:
          type: ["string", "null"]
        createdAt:
          type: string
      required:
        - id
        - title
        - url
        - fileType
        - targetVoicePart
        - createdAt
    CreateMusicBody:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
        fileType:
          type: string
          enum: [pdf, mp3, other]
        targetVoicePart:
          type: ["string", "null"]
      required:
        - title
        - url
        - fileType
        - targetVoicePart
...