# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teenagers Choir API
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login with username and password
 */
export const LoginBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  role: zod.string(),
  voicePart: zod.string(),
});

/**
 * @summary Register as a new member
 */
export const RegisterBody = zod.object({
  username: zod.string(),
  password: zod.string(),
});

/**
 * @summary Logout current user
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary Get current authenticated user
 */
export const GetCurrentUserResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  role: zod.string(),
  voicePart: zod.string(),
});

/**
 * @summary Change own password
 */
export const ChangePasswordBody = zod.object({
  currentPassword: zod.string(),
  newPassword: zod.string(),
});

export const ChangePasswordResponse = zod.object({
  message: zod.string(),
});

/**
 * @summary List all members (admin only)
 */
export const ListUsersResponseItem = zod.object({
  id: zod.number(),
  username: zod.string(),
  role: zod.string(),
  voicePart: zod.string(),
  createdAt: zod.string(),
});
export const ListUsersResponse = zod.array(ListUsersResponseItem);

/**
 * @summary Update a member's voice part (admin only)
 */
export const UpdateVoicePartParams = zod.object({
  id: zod.coerce.number(),
});

export const UpdateVoicePartBody = zod.object({
  voicePart: zod.enum(["Soprano", "Alto", "Normal"]),
});

export const UpdateVoicePartResponse = zod.object({
  id: zod.number(),
  username: zod.string(),
  role: zod.string(),
  voicePart: zod.string(),
  createdAt: zod.string(),
});

/**
 * @summary Get today's attendance records (admin only)
 */
export const ListAttendanceResponseItem = zod.object({
  id: zod.number(),
  userId: zod.number(),
  username: zod.string(),
  voicePart: zod.string(),
  status: zod.enum(["present", "absent"]),
  date: zod.string(),
});
export const ListAttendanceResponse = zod.array(ListAttendanceResponseItem);

/**
 * @summary Toggle attendance for a user (admin only)
 */
export const UpdateAttendanceParams = zod.object({
  userId: zod.coerce.number(),
});

export const UpdateAttendanceBody = zod.object({
  status: zod.enum(["present", "absent"]),
});

export const UpdateAttendanceResponse = zod.object({
  id: zod.number(),
  userId: zod.number(),
  username: zod.string(),
  voicePart: zod.string(),
  status: zod.enum(["present", "absent"]),
  date: zod.string(),
});

/**
 * @summary Get current member's attendance stats
 */
export const GetMyAttendanceStatsResponse = zod.object({
  totalSessions: zod.number(),
  presentCount: zod.number(),
  absentCount: zod.number(),
  attendanceRate: zod.number(),
});

/**
 * @summary Get messages for the current user's voice part
 */
export const ListMessagesResponseItem = zod.object({
  id: zod.number(),
  content: zod.string(),
  targetVoicePart: zod.string().nullable(),
  createdAt: zod.string(),
  authorUsername: zod.string(),
});
export const ListMessagesResponse = zod.array(ListMessagesResponseItem);

/**
 * @summary Send a broadcast message (admin only)
 */
export const CreateMessageBody = zod.object({
  content: zod.string(),
  targetVoicePart: zod.string().nullable(),
});

/**
 * @summary Delete a message (admin only)
 */
export const DeleteMessageParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get music files for current user's voice part
 */
export const ListMusicResponseItem = zod.object({
  id: zod.number(),
  title: zod.string(),
  url: zod.string(),
  fileType: zod.enum(["pdf", "mp3", "other"]),
  targetVoicePart: zod.string().nullable(),
  createdAt: zod.string(),
});
export const ListMusicResponse = zod.array(ListMusicResponseItem);

/**
 * @summary Upload a music link (admin only)
 */
export const CreateMusicBody = zod.object({
  title: zod.string(),
  url: zod.string(),
  fileType: zod.enum(["pdf", "mp3", "other"]),
  targetVoicePart: zod.string().nullable(),
});

/**
 * @summary Delete a music file (admin only)
 */
export const DeleteMusicParams = zod.object({
  id: zod.coerce.number(),
});

/**
 * @summary Get admin dashboard statistics
 */
export const GetAdminStatsResponse = zod.object({
  totalMembers: zod.number(),
  sopranoCount: zod.number(),
  altoCount: zod.number(),
  normalCount: zod.number(),
  presentToday: zod.number(),
  absentToday: zod.number(),
  totalMessages: zod.number(),
  totalMusicFiles: zod.number(),
});