# lib/api-zod/src/generated/types/createMessageBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teenagers Choir API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMessageBody {
  content: string;
  /** @nullable */
  targetVoicePart: string | null;
}