# lib/api-zod/src/generated/types/createMusicBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teenagers Choir API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateMusicBodyFileType } from "./createMusicBodyFileType";

export interface CreateMusicBody {
  title: string;
  url: string;
  fileType: CreateMusicBodyFileType;
  /** @nullable */
  targetVoicePart: string | null;
}