# lib/api-zod/src/generated/types/message.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teenagers Choir API
 * OpenAPI spec version: 0.1.0
 */

export interface Message {
  id: number;
  content: string;
  /** @nullable */
  targetVoicePart: string | null;
  createdAt: string;
  authorUsername: string;
}