# lib/api-zod/src/generated/types/musicFile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Teenagers Choir API
 * OpenAPI spec version: 0.1.0
 */
import type { MusicFileFileType } from "./musicFileFileType";

export interface MusicFile {
  id: number;
  title: string;
  url: string;
  fileType: MusicFileFileType;
  /** @nullable */
  targetVoicePart: string | null;
  createdAt: string;
}